+++ UnityUtils/Assets/Testing Scripts/TestingScript.cs
@@ -14,8 +14,6 @@ public class TestingScript : MonoBehaviour {
     ObjectAnimations.AnimateTransformScale(transform, transform.localScale, new Vector3(3, 4, 5), 2);
     ObjectAnimations.AnimateTransformPosition(transform, transform.position, new Vector3(0, 5, 0), 2, animationCurve: curve);
 
-    ObjectAnimations.AnimateSpriteRendererOpacity(GetComponent<SpriteRenderer>(), 0.5f, 1, 2);
-
     SFX test = new SFX(clip).SetPitch(3, 4);
   }
 }