+++ UnityUtils/Packages/com.mydimons.unityutils/CHANGELOG.md
@@ -2,9 +2,11 @@
 
 ### Additions
 - SFX class to play audio clips with set parameters
+- ColorFlash class to play color flashes with set parameters
 
 ### Changes
 - SFXManager.cs now only accepts SFX objects as parameters (Removed all other functions)
+- ColorFlashManager.cs now uses its ColorFlash colorFlash as the default Flash() parameter
 - AudioManager.cs CalculatePitchVariance() now takes in a starting pitch.
 - ObjectColorFlash.cs renamed to "ColorFlashManager.cs"
 - Organized ObjectModifiers into their own Modifiers namespace